scripts: dnm: add script to enforce "Do Not Merge" labels in PRs
This patch introduces a Python script that checks if any labels defined
in DNM_LABELS are present on a pull request. If a label is found,
the script exits with status 1, causing the associated GitHub Actions
workflow to fail. Otherwise, it exits with status 0.

This script is intended to be used only within GitHub Actions workflows.

This script was obtained from commit
649eb6280da86f984b96be92d456472eb6ac75c1
of https://github.com/zephyrproject-rtos/zephyr
and has been modified very slightly.

Signed-off-by: Mirai SHINJO <oss@mshinjo.com>