boards: add empty overlay files for Nucleo boards
Add empty overlay files for Nucleo boards. These overlays will
be updated as necessary to align peripheral availability
between in-house designed boards and development kits.

This is important because in-house designed boards require
separate power sources, which is inconvenient for software
development, and their availability in terms of quantity
and development cycles is not guaranteed.

Signed-off-by: David Jin <youyouare123@gmail.com>