ccsds123b: introduce a benchmark test
A board with a flashed benchmark immediately starts compressing the
image upon boot, and then prints the time it took to finish the
compression.

The benchmark test reimplements arrays.c and image.c to support testing
on an image of any size. A larger image is generated by tiling a
smaller one.

Signed-off-by: Igor Chovpan <i.chovpan@gmail.com>