mcuboot: adding bootloader and blinky1/2 swap demo
Adding a full mcuboot that swaps between blinky1.c and blinky2.c
which are two basic blinky images that blink at different
frequencies. On a button press, the bootloader will note that on
next reboot, the other swap image should be used. Config files
might need to change later. Using ecdsa as the signing for the
images.

Signed-off-by: Juan Pablo Medina <jpmedinagl@gmail.com>